+++ index.js
@@ -88,6 +88,12 @@ var spawn = function(opts) {
   child.stdout.pipe(debugStream('phantom (%s) stdout', child.pid)).pipe(output);
   input.pipe(debugStream('phantom (%s) stdin', child.pid)).pipe(child.stdin);
 
+  // Phantom may print to STDERR if the --debug flag is used.
+  // Pass through Phantom's STDERR to our own, while adding a prefix so we can tell them apart.
+  child.stderr.on('data', function (data) {
+    console.error("[phantom stderr] "+data);
+  });
+
   var onerror = once(function() {
     child.kill();
   });